Moved ErrorResponse to shared folder
